+++ tests/models/test_ngsi_v2_context.py
@@ -5,14 +5,14 @@ import unittest
 from typing import List
 
 from pydantic import ValidationError
+
+from filip.models.ngsi_v2.base import Metadata, NamedMetadata
 from filip.models.ngsi_v2.context import \
     ActionType, \
     Command, \
-    ContextMetadata, \
     ContextAttribute, \
     ContextEntity, \
     create_context_entity_model, \
-    NamedContextMetadata, \
     Update, \
     NamedContextAttribute, \
     ContextEntityKeyValues, \
@@ -61,9 +61,9 @@ class TestContextModels(unittest.TestCase):
         Returns:
             None
         """
-        md1 = ContextMetadata(type='Text', value='test')
-        md2 = NamedContextMetadata(name='info', type='Text', value='test')
-        md3 = [NamedContextMetadata(name='info', type='Text', value='test')]
+        md1 = Metadata(type='Text', value='test')
+        md2 = NamedMetadata(name='info', type='Text', value='test')
+        md3 = [NamedMetadata(name='info', type='Text', value='test')]
         attr1 = ContextAttribute(value=20,
                                  type='Integer',
                                  metadata={'info': md1})
@@ -98,7 +98,7 @@ class TestContextModels(unittest.TestCase):
             exclude={'name', 'metadata'}, exclude_unset=True)})
 
         new_attr = {'new_attr': ContextAttribute(type='Number', value=25)}
-        entity.add_properties(new_attr)
+        entity.add_attributes(new_attr)
 
         generated_model = create_context_entity_model(data=self.entity_data)
         entity = generated_model(**self.entity_data)
@@ -152,9 +152,9 @@ class TestContextModels(unittest.TestCase):
         # Test if all needed fields, detect all invalid strings
         for string in invalid_strings:
             self.assertRaises(ValidationError,
-                              ContextMetadata, type=string)
+                              Metadata, type=string)
             self.assertRaises(ValidationError,
-                              NamedContextMetadata, name=string)
+                              NamedMetadata, name=string)
             self.assertRaises(ValidationError,
                               ContextAttribute, type=string)
             self.assertRaises(ValidationError,
@@ -168,8 +168,8 @@ class TestContextModels(unittest.TestCase):
 
         # Test if all needed fields, do not trow wrong errors
         for string in valid_strings:
-            ContextMetadata(type=string)
-            NamedContextMetadata(name=string)
+            Metadata(type=string)
+            NamedMetadata(name=string)
             ContextAttribute(type=string)
             NamedContextAttribute(name=string)
             ContextEntityKeyValues(id=string, type=string)
@@ -177,15 +177,15 @@ class TestContextModels(unittest.TestCase):
 
         # Test for the special-string protected field if all strings are blocked
         for string in special_strings:
+            self.assertRaises(ValidationError, ContextAttribute, type=string)
             self.assertRaises(ValidationError,
                               NamedContextAttribute, name=string)
             self.assertRaises(ValidationError,
                               NamedCommand, name=string)
         # Test for the normal protected field if all strings are allowed
         for string in special_strings:
-            ContextMetadata(type=string)
-            NamedContextMetadata(name=string)
-            ContextAttribute(type=string)
+            Metadata(type=string)
+            NamedMetadata(name=string)
             ContextEntityKeyValues(id=string, type=string)
 
     def tearDown(self) -> None: